refactor: 코드 리뷰 반영 - RecuritCard 컴포넌트의 순환 참조를 수정 했습니다. - RecuirtCard 컴포넌트의 icon 경로를 수정했습니다. - route path를 수정했습니다.
